Architecture update, refactor, 82% accuracy on MNIST classification
